Fully updated to 1.8
* Dropped support for 1.7 entirely
* Fixed player, primed TNT, fireworks and item frame disguises
* Removed experimental future disguise types feature

+++ src/me/libraryaddict/disguise/disguisetypes/watchers/EndermanWatcher.java
@@ -36,7 +36,7 @@ public class EndermanWatcher extends LivingWatcher {
 
     @Override
     public void setItemInHand(ItemStack itemstack) {
-        setValue(16, (byte) (itemstack.getTypeId() & 255));
+        setValue(16, (short) (itemstack.getTypeId() & 255));
         setValue(17, (byte) (itemstack.getDurability() & 255));
     }
